# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Exotic India - MIS",
    "description": "API for managing simple tasks",
    "version": "1.0.0"
  },
  "host": "localhost:4000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "x-api-key",
      "description": "API Key authentication. Pass your secret key in the x-api-key header."
    }
  },
  "paths": {
    "/metrics": {
      "get": {
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/api/runQuery": {
      "post": {
        "tags": [
          "General Purpose"
        ],
        "description": "Enter Any Query in Request body it will provide json response",
        "parameters": [
          {
            "name": "x-forwarded-for",
            "in": "header",
            "type": "string"
          },
          {
            "name": "x-api-key",
            "in": "header",
            "type": "string"
          },
          {
            "name": "api-key",
            "in": "header",
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "api_key",
            "in": "query",
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "description": "SQL QUERY",
            "schema": {
              "type": "object",
              "properties": {
                "sql": {
                  "type": "string",
                  "example": "select * from products limit 1 "
                }
              },
              "required": [
                "sql"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/api/users": {
      "get": {
        "tags": [
          "User Management"
        ],
        "description": "List all authorized API users and their API keys",
        "parameters": [
          {
            "name": "x-forwarded-for",
            "in": "header",
            "type": "string"
          },
          {
            "name": "x-api-key",
            "in": "header",
            "type": "string"
          },
          {
            "name": "api-key",
            "in": "header",
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "api_key",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "User Management"
        ],
        "description": "Create a new user and generate an API key for them",
        "parameters": [
          {
            "name": "x-forwarded-for",
            "in": "header",
            "type": "string"
          },
          {
            "name": "x-api-key",
            "in": "header",
            "type": "string"
          },
          {
            "name": "api-key",
            "in": "header",
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "api_key",
            "in": "query",
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "description": "User details",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string",
                  "example": "john"
                },
                "role": {
                  "type": "string",
                  "example": "user"
                },
                "customApiKey": {
                  "type": "string",
                  "example": "optional-custom-key"
                }
              },
              "required": [
                "username"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/api/users/{username}": {
      "delete": {
        "tags": [
          "User Management"
        ],
        "description": "Delete a user by username",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "x-forwarded-for",
            "in": "header",
            "type": "string"
          },
          {
            "name": "x-api-key",
            "in": "header",
            "type": "string"
          },
          {
            "name": "api-key",
            "in": "header",
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "api_key",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    }
  },
  "security": [
    {
      "apiKeyAuth": []
    }
  ]
}